Add a logger aspect config.
LoggerConfig, for the moment, will consist of name for the logs collection and a list of descriptor names for log entries.

The log_entry_descriptor is also updated to include metadata about a special payload attribute. Non-payload attributes will be considered metadata attributes about the detail provided in the payload.