Simplify protogen container builder
The old continer builder was based on Golang 1.9 (according to
@geeknoid; I couldn't actually get it to build).  The container
could use some serious simplificaiton.  If we don't need to pin gRPC
and protobuf, it is better practice to let the experts in golang
distribute their own container (golang:1.12) and work from that.

I have assumed in this PR that they have the correct versions of
gRPC and protobuf, which may be a bad assumption.

Also this PR does not yet build (so don't merge) because of a
dependency problem in the tools repo.  Further, this PR likely
won't build and certainly isn't tested yet, until the tools
repo build problem is sorted out.

Depends-On: https://github.com/istio/tools/pull/222