VirtualService: use a double for mirror_percent (#1173)
To support mirroring less than 1% of traffic to a destination, the
mirror_percent field should provide specification of percentages less
than 1. Envoy permits mirroring fractions as small as 1 in 1,000,000.
The Delay and Abort messages already encountered this issue and
solved it by using a Percent (== double) field called percentage instead
of an integer field called percent.

Here, we support small fractions in mirroring using the same pattern,
using mirror_percentage in place of mirror_percent and deprecating the
mirror_percent field.

When reading mirror_percentage, we will need to handle rounding
of numbers near 1.0e-6 carefully because the floating point
representation will not be exact.