Remove redundant connect timeout (#1352)
This field exists in both mesh config and proxy config. The one I am
removing is in proxy config, as it is very confusing to have the same
field in two different places. This is currently used in one single
place: XDS cluster timeout:
https://github.com/istio/istio/blob/d131b957ac4af0e92fd480657198fe519ec7fc43/tools/packaging/common/envoy_bootstrap_v2.json#L308

I don't see a strong need to have that configurable, but if we do we can
just pull it from the mesh config setting.