tracing: add API for OpenCensus agent exporter (#1563)
* tracing: add API for OpenCensus agent exporter

This is based heavily on the existing Stackdriver configuration.
Stackdriver and OpenCensus exporter will both use the OpenCensus tracer
implementation. This will permit OpenCensus to export OpenCensus
agent-formatted spans.

OpenCensus-formatted spans are handled by the OpenTelemetry collector,
providing a migration path from OpenCensus to OpenTelemetry when
OpenTelemetry is more complete and integrated with Envoy.

Change-Id: I566299c9291021d26ec1e839643871a380807bc3
Reviewed-on: https://gerrit.musta.ch/c/public/istio-api/+/551
Reviewed-by: Jungho Ahn <jungho.ahn@airbnb.com>

* tracing: add TraceContext & use OpenCensusAgent

Specify the set of trace headers to use for distributed trace context
propagation with a new TraceContext enum in the OpenCensusAgent tracer.

OpenCensus is the name of both the tracer implementation and the backend
service. To disambiguate between the two terms, rename OpenCensus to
OpenCensusAgent, since it is specific to the backend service. Both
Stackdriver & OpenCensus will use the OpenCensus trace implementation
under the hood.

Change-Id: Id860fbf6bf6b1345c7276bb2c2b93de3f69a9d44
Reviewed-on: https://gerrit.musta.ch/c/public/istio-api/+/575
Reviewed-by: Jungho Ahn <jungho.ahn@airbnb.com>

* tracing: update docs to include envoy proto link

The existing documentation for Stackdriver and OpenCensus reference the
OpenCensus configuration. We should reference both the Envoy OpenCensus
tracer configuration and the OpenCensus configuration.

Change-Id: Ia3fcbc879e2e15f90a56959cf97fe30b57211782
Reviewed-on: https://gerrit.musta.ch/c/public/istio-api/+/582
Reviewed-by: Ying Zhu <ying.zhu@airbnb.com>
Reviewed-by: Jungho Ahn <jungho.ahn@airbnb.com>