More updates to purge proto - may not be right.
This PR changes the proto.lock file. I don't think that
is the right approach. We may be stuck with gogo in our APIs
until the end of time, unless there is some way to properly reserve
the messages that gogo consumes. THe other part of this work is in
the istio/istio repo, where I have started code to make the Model.Config
struct marshal/unmarshal gogo and proto into proto. This is mandatory
because K8s continues to use gogo...