Add topologySpreadConstraints field to operator k8s opts
This adds the topologySpreadConstraints field to the k8s section of the
operator CRD.

This field is an alternative to common anti-affinity configurations that attempt
to configure pods to spread evenly across nodes and availability zones.
This is important for running highly available services, and is
something that people will want to do with their Istio resources.