Running proto-lock (#2230)
As per https://github.com/istio/api#updating `make proto-lock`
has to be run as part of new api changes. I was trying
to run the same for my PR https://github.com/istio/api/pull/2207
and saw some additional changes added in proto.lock which
was not introduced by my change. Hence putting them in a separate
PR here to check if these changes are needed or not.

Signed-off-by: Faseela K <faseela.k@est.tech>