ambient upstream: bring mesh config to `master`
This is part of #40879. See that issue description for context.

This introduces a new configuration to enable/disable ambient mesh.
Enablement means that traffic is captured.

There are 3 modes: OFF, ON, and label selector.

To enable admins to globally enforce the mesh is enabled but skip a few
namespaces, a disable selector is present which restricts the scope of
ON mode.