Drop json schema generation (#2875)
We only generate this for some resources and not others, and no one has
complained -- so there doesn't seem to be much value in maintaining
this, but it provides a lot of boilerplate.

Rather than "fixing" it to be used for all types, seems better to just
drop this. We already have the schema as proto and OpenAPI (via CRD).