Add JSON marshaling for non-CRD types (#3345)
We have this for CRD types, where it is strictly required to work with
k8s, but not for our other types. This makes it problematic to marshal
json for these types.

Note you can directly use protobuf json library, but this doesn't work
for nested types, hence the desire to have these. Low cost, some value.